openissue-ai/backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.27.0
sentence-transformers>=2.2.0
faiss-cpu>=1.9.0
numpy>=1.26.0
pydantic>=2.0.0
python-multipart>=0.0.6
aiosqlite>=0.19.0
